Fix support for multiple networks
Instead of fixing the webhook handler to the network where the plugin was loaded, dynamically search through networks (much like the plugin already did with channels).

This also makes the plguin raise an explicit error if a network+channel target is unknown, instead of leaving any frontend servers with confusing 502 (Bad Gateway) errors.